chore: move src/utils/waitFor to tests/utils/
It's not used in production anymore, so move it out of src/
